pyproject.toml
[tool.poetry]
name = "fivethirtyone-db"
version = "0.1.0"
description = ""
authors = ["jkr <jimmy.kjaersgaard@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fivethirtyone_db"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "*"
mysql-connector-python = "^8.0.32"
ipywidgets = "^8.0.4"
panel = "^0.14.4"
pandas = "^1.5.3"
numpy = "^1.24.2"

[tool.poetry.group.dev.dependencies]
jupyterlab = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-sugar = "*"


[tool.pytest.ini_options]
testpaths = [
    "tests",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
